Specify module name to adapter on configuration
