+++ test/adapters/erlang_binary_diff_test.exs
@@ -1,7 +1,7 @@
-defmodule Adapters.ErlangBinary.DiffTest do
+defmodule Adapters.ErlangBinaryDiffTest do
   use ExUnit.Case
 
-  alias ExAudit.Adapters.ErlangBinary.Diff
+  alias ExAudit.Adapters.ErlangBinaryDiff, as: Diff
   doctest Diff
 
   test "should diff primitives" do
@@ -83,5 +83,37 @@ defmodule Adapters.ErlangBinary.DiffTest do
     b = %{foo: val2}
 
     assert %{foo: {:changed, {:primitive_change, val1, val2}}} == Diff.diff(a, b)
+  end
+
+  test "should apply primitive changes" do
+    assert_patching(:foo, :bar)
+  end
+
+  test "should cope with :not_changed" do
+    assert_patching(:foo, :foo)
+  end
+
+  test "should apply patches to lists" do
+    assert_patching([1, 2, 3], [1, 2, 4, 5])
+    assert_patching([1, 2, 3], [])
+  end
+
+  test "should apply patches to maps" do
+    assert_patching(%{a: 1, b: 2}, %{a: 3, b: 4})
+    assert_patching(%{}, %{foo: 3})
+    assert_patching(%{foo: 42}, %{})
+  end
+
+  test "should apply patches to complex structures" do
+    a = [%{foo: [1, 2, 3]}]
+    b = [%{foo: [1, 2, 3, 4]}]
+
+    assert_patching(a, b)
+  end
+
+  defp assert_patching(a, b) do
+    patch = Diff.diff(a, b)
+    assert b == Diff.patch(a, patch)
+    assert a == Diff.patch(b, Diff.reverse(patch))
   end
 end
