+++ test/adapters/map_diff_test.exs
@@ -1,7 +1,7 @@
-defmodule Adapters.Map.DiffTest do
+defmodule Adapters.MapDiffTest do
   use ExUnit.Case
 
-  alias ExAudit.Adapters.Map.Diff
+  alias ExAudit.Adapters.MapDiff, as: Diff
 
   test "should diff primitives" do
     assert %{added: :bar, changed: :primitive_change, removed: :foo} = Diff.diff(:foo, :bar)
@@ -102,5 +102,37 @@ defmodule Adapters.Map.DiffTest do
                changed: %{added: val2, changed: :primitive_change, removed: val1}
              }
            } == Diff.diff(a, b)
+  end
+
+  test "should apply primitive changes" do
+    assert_patching(:foo, :bar)
+  end
+
+  test "should cope with :not_changed" do
+    assert_patching(:foo, :foo)
+  end
+
+  test "should apply patches to lists" do
+    assert_patching([1, 2, 3], [1, 2, 4, 5])
+    assert_patching([1, 2, 3], [])
+  end
+
+  test "should apply patches to maps" do
+    assert_patching(%{a: 1, b: 2}, %{a: 3, b: 4})
+    assert_patching(%{}, %{foo: 3})
+    assert_patching(%{foo: 42}, %{})
+  end
+
+  test "should apply patches to complex structures" do
+    a = [%{foo: [1, 2, 3]}]
+    b = [%{foo: [1, 2, 3, 4]}]
+
+    assert_patching(a, b)
+  end
+
+  defp assert_patching(a, b) do
+    patch = Diff.diff(a, b)
+    assert b == Diff.patch(a, patch)
+    assert a == Diff.patch(b, Diff.reverse(patch))
   end
 end
